BUG: Fix "index out of bounds" error
The final step of selecting the maximum values to the pandas DataFrame
fails with: "Applying IndexError: positional indexers are
out-of-bounds".

This was because pandas preserves the original indices when filtering.
Typically one would reindex, but there's no need since the dataset is
only about 100 MB one may as well keep all the data in memory and
reference it.